move UpdateActionProps interface to a types file
