+++ app/client/src/workers/Evaluation/handlers/types.ts
@@ -1,0 +1,6 @@
+export interface UpdateActionProps {
+  entityName: string;
+  dataPath: string;
+  data: unknown;
+  dataPathRef?: string;
+}
